+++ pkg/cli/clisqlclient/make_query.go
@@ -12,8 +12,10 @@ package clisqlclient
 
 import (
 	"database/sql/driver"
+	"strings"
 
 	"github.com/cockroachdb/cockroach/pkg/sql/parser"
+	"github.com/cockroachdb/errors"
 )
 
 // QueryFn is the type of functions produced by MakeQuery.
@@ -38,6 +40,26 @@ func MakeQuery(query string, parameters ...driver.Value) QueryFn {
 			}
 		}
 		rows, err := conn.Query(query, parameters)
+		err = handleCopyError(conn.(*sqlConn), err)
 		return rows, isMultiStatementQuery, err
 	}
+}
+
+// handleCopyError ensures the user is properly informed when they issue
+// a COPY statement somewhere in their input.
+func handleCopyError(conn *sqlConn, err error) error {
+	if err == nil {
+		return nil
+	}
+
+	if !strings.HasPrefix(err.Error(), "pq: unknown response for simple query: 'G'") {
+		return err
+	}
+
+	// The COPY statement has hosed the connection by putting the
+	// protocol in a state that lib/pq cannot understand any more. Reset
+	// it.
+	_ = conn.Close()
+	conn.reconnecting = true
+	return errors.New("woops! COPY has confused this client! Suggestion: use 'psql' for COPY")
 }
